+++ src/probnum/filtsmooth/gaussfiltsmooth/extendedkalman.py
@@ -2,109 +2,51 @@
 Gaussian filtering and smoothing based on making intractable quantities
 tractable through Taylor-method approximations, e.g. linearization.
 """
-import numpy as np
 
-from probnum.filtsmooth.gaussfiltsmooth._utils import is_cont_disc, is_disc_disc
-from probnum.filtsmooth.gaussfiltsmooth.gaussfiltsmooth import (
-    GaussFiltSmooth,
-    linear_discrete_update,
-)
-from probnum.filtsmooth.statespace import DiscreteGaussianModel, LinearSDEModel
+from probnum.filtsmooth import statespace
 from probnum.random_variables import Normal
 
 
-class ExtendedKalman(GaussFiltSmooth):
-    """
-    Factory method for Kalman filters.
-    """
+class ContinuousEKF(statespace.Transition):
+    """Continuous extended Kalman filter transition."""
+    def __init__(self, cont_model):
+        if not isinstance(cont_model, statespace.ContinuousModel):
+            raise TypeError
+        self.cont_model = cont_model
 
-    def __new__(cls, dynamod, measmod, initrv, **kwargs):
+    def transition_realization(self, real, start, stop, **kwargs):
+        return self.cont_model.transition_realization(real, start, stop, **kwargs)
 
-        if cls is ExtendedKalman:
-            if is_cont_disc(dynamod, measmod):
-                return _ContDiscExtendedKalman(dynamod, measmod, initrv, **kwargs)
-            if is_disc_disc(dynamod, measmod):
-                return _DiscDiscExtendedKalman(dynamod, measmod, initrv, **kwargs)
-            else:
-                errmsg = (
-                    "Cannot instantiate Extended Kalman filter with "
-                    "given dynamic model and measurement model."
-                )
-                raise ValueError(errmsg)
-        else:
-            return super().__new__(cls)
+    def transition_rv(self, rv, start, stop, **kwargs):
+        raise NotImplementedError
+
+    @property
+    def dimension(self):
+        raise NotImplementedError
 
 
-class _ContDiscExtendedKalman(ExtendedKalman):
-    """
-    Continuous-discrete extended Kalman filtering and smoothing.
-    """
+class DiscreteEKF(statespace.Transition):
+    """Discrete extended Kalman filter transition."""
+    def __init__(self, disc_model):
+        self.disc_model = disc_model
 
-    def __init__(self, dynamod, measmod, initrv, **kwargs):
-        if not issubclass(type(dynamod), LinearSDEModel):
-            raise ValueError(
-                "This implementation of ContDiscExtendedKalman "
-                "requires a linear dynamic model."
-            )
-        if not issubclass(type(measmod), DiscreteGaussianModel):
-            raise ValueError(
-                "ContDiscExtendedKalman requires a Gaussian measurement model."
-            )
-        if "cke_nsteps" in kwargs.keys():
-            self.cke_nsteps = kwargs["cke_nsteps"]
-        else:
-            self.cke_nsteps = 1
-        super().__init__(dynamod, measmod, initrv)
+    def transition_realization(self, real, start, stop, **kwargs):
+        return self.disc_model.transition_realization(real, start, stop, **kwargs)
 
-    def predict(self, start, stop, randvar, **kwargs):
-        step = (stop - start) / self.cke_nsteps
-        return self.dynamicmodel.transition_rv(
-            rv=randvar, start=start, stop=stop, step=step, **kwargs
-        )
-
-    def update(self, time, randvar, data, **kwargs):
-        return _discrete_extkalman_update(
-            time, randvar, data, self.measurementmodel, **kwargs
-        )
-
-
-class _DiscDiscExtendedKalman(ExtendedKalman):
-    def __init__(self, dynamod, measmod, initrv, **kwargs):
-        """
-        Checks that dynamod and measmod are linear and moves on.
-        """
-        if not issubclass(type(dynamod), DiscreteGaussianModel):
-            raise ValueError(
-                "DiscDiscExtendedKalmanFilter requires " "a linear dynamic model."
-            )
-        if not issubclass(type(measmod), DiscreteGaussianModel):
-            raise ValueError(
-                "DiscDiscExtendedKalmanFilter requires " "a linear measurement model."
-            )
-        super().__init__(dynamod, measmod, initrv)
-
-    def predict(self, start, stop, randvar, **kwargs):
-        mean, covar = randvar.mean, randvar.cov
-        if np.isscalar(mean) and np.isscalar(covar):
-            mean, covar = mean * np.ones(1), covar * np.eye(1)
-        diffmat = self.dynamod.diffusionmatrix(start, **kwargs)
-        jacob = self.dynamod.jacobian(start, mean, **kwargs)
-        mpred = self.dynamod.dynamics(start, mean, **kwargs)
-        crosscov = covar @ jacob.T
+    def transition_rv(self, rv, start, **kwargs):
+        diffmat = self.disc_model.diffusionmatrix(start)
+        jacob = self.disc_model.jacobian(start, rv.mean)
+        mpred = self.disc_model.dynamics(start, rv.mean)
+        crosscov = rv.cov @ jacob.T
         cpred = jacob @ crosscov + diffmat
         return Normal(mpred, cpred), {"crosscov": crosscov}
 
-    def update(self, time, randvar, data, **kwargs):
-        return _discrete_extkalman_update(
-            time, randvar, data, self.measurementmodel, **kwargs
-        )
+    @property
+    def dimension(self):
+        raise NotImplementedError
 
+    @staticmethod
+    def from_ode(self, ode, integrator):
+        """Will replace `ivp2ekf` soon... """
+        raise NotImplementedError
 
-def _discrete_extkalman_update(time, randvar, data, measmod, **kwargs):
-    mpred, cpred = randvar.mean, randvar.cov
-    if np.isscalar(mpred) and np.isscalar(cpred):
-        mpred, cpred = mpred * np.ones(1), cpred * np.eye(1)
-    jacob = measmod.jacobian(time, mpred, **kwargs)
-    meascov = measmod.diffusionmatrix(time, **kwargs)
-    meanest = measmod.dynamics(time, mpred, **kwargs)
-    return linear_discrete_update(meanest, cpred, data, meascov, jacob, mpred)
